Moving remaining datetime functions from bigquery helpers to core.
