+++ gcloud/bigquery/table.py
@@ -18,9 +18,9 @@ import datetime
 
 import six
 
+from gcloud._helpers import _datetime_from_prop
+from gcloud._helpers import _prop_from_datetime
 from gcloud.exceptions import NotFound
-from gcloud.bigquery._helpers import _datetime_from_prop
-from gcloud.bigquery._helpers import _prop_from_datetime
 
 
 _MARKER = object()
