+++ gcloud/test__helpers.py
@@ -300,6 +300,80 @@ class Test__millis(unittest2.TestCase):
         self.assertEqual(self._callFUT(WHEN), 1000)
 
 
+class Test__datetime_from_prop(unittest2.TestCase):
+
+    def _callFUT(self, value):
+        from gcloud._helpers import _datetime_from_prop
+        return _datetime_from_prop(value)
+
+    def test_w_none(self):
+        self.assertTrue(self._callFUT(None) is None)
+
+    def test_w_millis(self):
+        import datetime
+        from gcloud._helpers import UTC
+        from gcloud._helpers import _TOTAL_SECONDS
+
+        NOW = datetime.datetime(2015, 7, 29, 17, 45, 21, 123456,
+                                tzinfo=UTC)
+        EPOCH = datetime.datetime(1970, 1, 1, tzinfo=UTC)
+        MILLIS = _TOTAL_SECONDS(NOW - EPOCH) * 1000
+        self.assertEqual(self._callFUT(MILLIS), NOW)
+
+
+class Test__prop_from_datetime(unittest2.TestCase):
+
+    def _callFUT(self, value):
+        from gcloud._helpers import _prop_from_datetime
+        return _prop_from_datetime(value)
+
+    def test_w_none(self):
+        self.assertTrue(self._callFUT(None) is None)
+
+    def test_w_utc_datetime(self):
+        import datetime
+        from gcloud._helpers import UTC
+        from gcloud._helpers import _TOTAL_SECONDS
+
+        NOW = datetime.datetime.utcnow().replace(tzinfo=UTC)
+        EPOCH = datetime.datetime(1970, 1, 1, tzinfo=UTC)
+        MILLIS = int(_TOTAL_SECONDS(NOW - EPOCH) * 1000)
+        result = self._callFUT(NOW)
+        self.assertTrue(isinstance(result, int))
+        self.assertEqual(result, MILLIS)
+
+    def test_w_non_utc_datetime(self):
+        import datetime
+        from gcloud._helpers import UTC
+        from gcloud._helpers import _UTC
+        from gcloud._helpers import _TOTAL_SECONDS
+
+        class CET(_UTC):
+            _tzname = 'CET'
+            _utcoffset = datetime.timedelta(hours=-1)
+
+        zone = CET()
+        NOW = datetime.datetime(2015, 7, 28, 16, 34, 47, tzinfo=zone)
+        EPOCH = datetime.datetime(1970, 1, 1, tzinfo=UTC)
+        MILLIS = int(_TOTAL_SECONDS(NOW - EPOCH) * 1000)
+        result = self._callFUT(NOW)
+        self.assertTrue(isinstance(result, int))
+        self.assertEqual(result, MILLIS)
+
+    def test_w_naive_datetime(self):
+        import datetime
+        from gcloud._helpers import UTC
+        from gcloud._helpers import _TOTAL_SECONDS
+
+        NOW = datetime.datetime.utcnow()
+        UTC_NOW = NOW.replace(tzinfo=UTC)
+        EPOCH = datetime.datetime(1970, 1, 1, tzinfo=UTC)
+        MILLIS = int(_TOTAL_SECONDS(UTC_NOW - EPOCH) * 1000)
+        result = self._callFUT(NOW)
+        self.assertTrue(isinstance(result, int))
+        self.assertEqual(result, MILLIS)
+
+
 class _AppIdentity(object):
 
     def __init__(self, app_id):
